Create an ROC curve from observed data

Create an ROC curve from observed data

Usage

compute_empirical_roc(
  data,
  response,
  predictor,
  direction = "auto",
  best_weights = c(1, 0.5),
  ...
)

Arguments

data

a dataframe containing responses (groupings) and predictor variable

response

a bare column name with the group status (control vs. cases)

predictor

a bare column name with the predictor to use for classification

direction

direction to set for the for pROC::roc(). Defaults to "auto".

best_weights

weights for computing the best ROC curve points. Defaults to c(1, .5), which are the defaults used by pROC::coords().

...

additional arguments passed to pROC::roc().

Value

a new dataframe of ROC coordinates is returned with columns for the predictor variable, .sensitivities, .specificities, .auc, .direction, .controls, .cases, .is_best_youden and .is_best_closest_topleft.

Examples

set.seed(100)
x1 <- rnorm(100, 4, 1)
x2 <- rnorm(100, 2, .5)
both <- c(x1, x2)
steps <- seq(min(both), max(both), length.out = 200)
d1 <- dnorm(steps, mean(x1), sd(x1))
d2 <- dnorm(steps, mean(x2), sd(x2))
data <- tibble::tibble(
  y = steps,
  d1 = d1,
  d2 = d2,
  outcome = rbinom(200, 1, prob = 1 - (d1 / (d1 + d2))),
  group = ifelse(outcome, "case", "control")
)

# get an ROC on the fake data
compute_empirical_roc(data, outcome, y)
#> Setting levels: control = 0, case = 1
#> Setting direction: controls > cases
#> # A tibble: 201 × 9
#>         y .specificities .sensitivities  .auc .direction .controls .cases
#>     <dbl>          <dbl>          <dbl> <dbl> <chr>      <chr>     <chr> 
#>  1 Inf           0                    1 0.979 >          0         1     
#>  2   6.57        0.00704              1 0.979 >          0         1     
#>  3   6.54        0.0141               1 0.979 >          0         1     
#>  4   6.51        0.0211               1 0.979 >          0         1     
#>  5   6.48        0.0282               1 0.979 >          0         1     
#>  6   6.45        0.0352               1 0.979 >          0         1     
#>  7   6.43        0.0423               1 0.979 >          0         1     
#>  8   6.40        0.0493               1 0.979 >          0         1     
#>  9   6.37        0.0563               1 0.979 >          0         1     
#> 10   6.34        0.0634               1 0.979 >          0         1     
#> # … with 191 more rows, and 2 more variables: .is_best_youden <lgl>,
#> #   .is_best_closest_topleft <lgl>
# this guess the cases and controls from the group name and gets it wrong
compute_empirical_roc(data, group, y)
#> Setting levels: control = case, case = control
#> Setting direction: controls < cases
#> # A tibble: 201 × 9
#>           y .specificities .sensitivities  .auc .direction .controls .cases 
#>       <dbl>          <dbl>          <dbl> <dbl> <chr>      <chr>     <chr>  
#>  1 -Inf             0                   1 0.979 <          case      control
#>  2    0.946         0.0172              1 0.979 <          case      control
#>  3    0.974         0.0345              1 0.979 <          case      control
#>  4    1.00          0.0517              1 0.979 <          case      control
#>  5    1.03          0.0690              1 0.979 <          case      control
#>  6    1.06          0.0862              1 0.979 <          case      control
#>  7    1.09          0.103               1 0.979 <          case      control
#>  8    1.12          0.121               1 0.979 <          case      control
#>  9    1.14          0.138               1 0.979 <          case      control
#> 10    1.17          0.155               1 0.979 <          case      control
#> # … with 191 more rows, and 2 more variables: .is_best_youden <lgl>,
#> #   .is_best_closest_topleft <lgl>
# better
compute_empirical_roc(data, group, y, levels = c("control", "case"))
#> Setting direction: controls > cases
#> # A tibble: 201 × 9
#>         y .specificities .sensitivities  .auc .direction .controls .cases
#>     <dbl>          <dbl>          <dbl> <dbl> <chr>      <chr>     <chr> 
#>  1 Inf           0                    1 0.979 >          control   case  
#>  2   6.57        0.00704              1 0.979 >          control   case  
#>  3   6.54        0.0141               1 0.979 >          control   case  
#>  4   6.51        0.0211               1 0.979 >          control   case  
#>  5   6.48        0.0282               1 0.979 >          control   case  
#>  6   6.45        0.0352               1 0.979 >          control   case  
#>  7   6.43        0.0423               1 0.979 >          control   case  
#>  8   6.40        0.0493               1 0.979 >          control   case  
#>  9   6.37        0.0563               1 0.979 >          control   case  
#> 10   6.34        0.0634               1 0.979 >          control   case  
#> # … with 191 more rows, and 2 more variables: .is_best_youden <lgl>,
#> #   .is_best_closest_topleft <lgl>