Items from the TOCS+ used by the WISC Lab

We use items from the Test of Children's Speech (TOCS+, Hodge & Gotzke, 2014) in our child speech and listener intelligibility studies.

Usage

data_tocs_items

Format

A data frame with 103 rows and 4 variables:

tocs_type

whether the item is a single-word or multiword utterance

tocs_level

number of words in the prompt

tocs_item

identifier for the item (e.g., within filenames)

tocs_prompt

the actual prompt presented to the child

tocs_item_note

any notes about an item's usage in experiments or analyses

These are the items:

tocs_type	tocs_level	tocs_item	tocs_prompt

multiword	2	S2T01	Bird house

multiword	2	S2T02	Cowboy boots

multiword	2	S2T03	Hug daddy

multiword	2	S2T04	Get off

multiword	2	S2T05	Animal crackers

multiword	2	S2T06	Living room

multiword	2	S2T07	Doctor bag

multiword	2	S2T08	Baby sock

multiword	2	S2T09	Big potato

multiword	2	S2T10	Say goodbye

multiword	3	S3T01	Gather the toys

multiword	3	S3T02	Five more cookies

multiword	3	S3T03	Make a birdhouse

multiword	3	S3T04	Tie those shoes

multiword	3	S3T05	Put these together

multiword	3	S3T06	That's not white

multiword	3	S3T07	Open the house

multiword	3	S3T08	Take his turn

multiword	3	S3T09	Pull his tooth

multiword	3	S3T10	Point to Teddy

multiword	3	S3T11	Cook big hotdogs

multiword	4	S4T01	She could hide here

multiword	4	S4T02	Wear a cowboy hat

multiword	4	S4T03	Find the brown lid

multiword	4	S4T04	Both faces are happy

multiword	4	S4T05	That is her dog

multiword	4	S4T06	He likes potato chips

multiword	4	S4T07	Jump over the box

multiword	4	S4T08	She chewed her fingers

multiword	4	S4T09	Cut two small pieces

multiword	4	S4T10	Get them some coffee

multiword	5	S5T01	Baby likes his new toy

multiword	5	S5T02	They'll eat those hotdogs soon

multiword	5	S5T03	His fingers are in wrong

multiword	5	S5T04	They are singing Happy Birthday

multiword	5	S5T05	Water shoots from that gun

multiword	5	S5T06	This cheese doesn't smell good

multiword	5	S5T07	The sign says keep out

multiword	5	S5T08	Tie up the garbage bag

multiword	5	S5T09	Give the flowers some water

multiword	5	S5T10	Put all the toys away

multiword	6	S6T01	He wants somebody to push him

multiword	6	S6T02	He gave Mommy a birthday card

multiword	6	S6T03	That baby is learning to talk

multiword	6	S6T04	She laughed at his funny jokes

multiword	6	S6T05	There's a playground near the school

multiword	6	S6T06	He winds up the toy ghost

multiword	6	S6T07	Make a check beside the two

multiword	6	S6T08	The cowboy hat has a feather

multiword	6	S6T09	Taste the cookies that she baked

multiword	6	S6T10	Take the lid off the pot

multiword	7	S7T01	Pour some tea for both of us

multiword	7	S7T02	The loud noise scared the cat away

multiword	7	S7T03	Don't let go or it will fall

multiword	7	S7T04	She is showing him how to jump

multiword	7	S7T05	Baby reached to get another chocolate cookie

multiword	7	S7T06	The bird landed beside the shaggy dog

multiword	7	S7T07	She got mad and pushed the boy

multiword	7	S7T08	They ate birthday cake and drank pop

multiword	7	S7T09	Be very quiet when Baby is sleeping

multiword	7	S7T10	Fill the fridge with things to eat

single-word	1	WT01	Pizza

single-word	1	WT02	Hug

single-word	1	WT03	Boot

single-word	1	WT04	Ball

single-word	1	WT05	Bee

single-word	1	WT06	Buddy

single-word	1	WT07	Hoot

single-word	1	WT08	Mud

single-word	1	WT09	A dress

single-word	1	WT10	Snow

single-word	1	WT11	Chew

single-word	1	WT12	Pan

single-word	1	WT13	Come

single-word	1	WT14	Boy

single-word	1	WT15	Bad

single-word	1	WT16	Hot

single-word	1	WT17	Gum

single-word	1	WT18	Dee

single-word	1	WT19	Beanie

single-word	1	WT20	Zoo

single-word	1	WT21	Hat

single-word	1	WT22	Seat

single-word	1	WT23	Yawn

single-word	1	WT24	No

single-word	1	WT25	Full

single-word	1	WT26	Bus

single-word	1	WT27	Jar

single-word	1	WT28	Top

single-word	1	WT29	Lock

single-word	1	WT30	Walk

single-word	1	WT31	Bow

single-word	1	WT32	Eat

single-word	1	WT33	Sheet

single-word	1	WT34	Rock

single-word	1	WT35	House

single-word	1	WT36	Pipe

single-word	1	WT37	Toys

single-word	1	WT38	Type

single-word	1	WT39	Wait

single-word	1	WT40	Whip

single-word	1	WT41	mIm

single-word	1	WT42	Big

These are the items with entries in the "tocs_item_note" fields:

tocs_type	tocs_level	tocs_item	tocs_prompt	tocs_item_note

multiword	3	S3T09	Pull his tooth	possibly not an original TOCS item

multiword	3	S3T11	Cook big hotdogs	practice item for intelligibility studies (usually not found in intelligibility data)

multiword	5	S5T10	Put all the toys away	practice item for intelligibility studies (usually not found in intelligibility data)

single-word	1	WT01	Pizza	practice item for intelligibility studies (usually not found in intelligibility data)

single-word	1	WT02	Hug	practice item for intelligibility studies (usually not found in intelligibility data)

single-word	1	WT09	A dress	exclude from intelligibility analyses for being two words and ambiguous with 'address'; possibly not an original TOCS item

single-word	1	WT18	Dee	sometimes exclude from intelligibility analyses for naming a letter (listeners are not inclined to treat a letter as a word); possibly not an original TOCS item

single-word	1	WT31	Bow	rhymes with 'snow'

single-word	1	WT35	House	noun; possibly not an original TOCS item

single-word	1	WT36	Pipe	possibly not an original TOCS item

single-word	1	WT37	Toys	possibly not an original TOCS item

single-word	1	WT38	Type	possibly not an original TOCS item

single-word	1	WT39	Wait	possibly not an original TOCS item

single-word	1	WT40	Whip	possibly not an original TOCS item

single-word	1	WT41	mIm	exclude from intelligibility analyses for being a nonword item; possibly not an original TOCS item

single-word	1	WT42	Big	possibly not an original TOCS item

Details

Most of these items are subset from a larger pool of items on the TOCS+. Hustad and colleagues (2021) is a supplemental material for a larger study about speech intelligibility which describes where the items were taken from within the larger TOCS+ task. (It is unclear, as I write this in 2026, which items were added by our group to the TOCS+ pool, so I have added "possibly not an original TOCS item" to items that do not appear in that supplemental material.)

In this picture-prompted repetition task, a child hears a prompt, repeats it (production), and later on listeners transcribe it (transcription). In our listener-response database, we differentiate these three forms with fields named tocs_prompt (prompt), sentence (production), and response (transcription).

References

Hodge, M. M., & Gotzke, C. (2014). Construct-related validity of the TOCS+ measures: Comparison of intelligibility and speaking rate scores in children with and without speech disorders. Journal of Communication Disorders, 51, 51–63. https://doi.org/10.1016/j.jcomdis.2014.06.007

Hustad, K. C., Mahr, T. J., Natzke, P., & Rathouz, P. J. (2021). Supplemental Material 1: Intelligibility growth between 30 and 119 months (Hustad et al., 2021) (Version 1). ASHA Journals. https://doi.org/10.23641/asha.16583426.v1