Fit a generalized gamma regression model (for speaking rate)

The function fits the same type of model as used in Mahr 2021: A generalized gamma regression model (via gamlss.dist::GG()) with natural cubic splines on the mean (mu), scale (sigma), and shape (nu) of the distribution. This model is fitted using this package's mem_gamlss() wrapper function.

Usage

fit_gen_gamma_gamlss(
  data,
  var_x,
  var_y,
  df_mu = 3,
  df_sigma = 2,
  df_nu = 1,
  control = NULL
)

fit_gen_gamma_gamlss_se(
  data,
  name_x,
  name_y,
  df_mu = 3,
  df_sigma = 2,
  df_nu = 1,
  control = NULL
)

predict_gen_gamma_gamlss(newdata, model, centiles = c(5, 10, 50, 90, 95))

Source

Associated article: https://doi.org/10.1044/2021_JSLHR-21-00206

Arguments

data

a data frame

var_x, var_y

(unquoted) variable names giving the predictor variable (e.g., age) and outcome variable (.e.g, rate).

df_mu, df_sigma, df_nu

degrees of freedom

control

a gamlss::gamlss.control() controller. Defaults to NULL which uses default settings, except for setting trace to FALSE to silence the output from gamlss.

name_x, name_y

quoted variable names giving the predictor variable (e.g., "age") and outcome variable (.e.g, "rate"). These arguments apply to fit_gen_gamma_gamlss_se().

newdata

a one-column dataframe for predictions

model

a model fitted by fit_gen_gamma_gamlss()

centiles

centiles to use for prediction. Defaults to c(5, 10, 50, 90, 95).

Value

for fit_gen_gamma_gamlss() and fit_gen_gamma_gamlss_se(), a mem_gamlss()-fitted model. The .user data in the model includes degrees of freedom for each parameter and the splines::ns() basis for each parameter. For predict_gen_gamma_gamlss(), a dataframe containing the model predictions for mu, sigma, and nu, plus columns for each centile in centiles.

Details

There are versions two of this function. The main version is fit_gen_gamma_gamlss(), and it works with unquoted column names (e.g., age). The alternative version is fit_gen_gamma_gamlss_se(); the final "se" stands for "Standard Evaluation". This designation means that the variable names must be given as strings (so, the quoted "age" instead of bare name age). This alternative version is necessary when we fit several models using parallel computing with furrr::future_map() (as when using bootstrap resampling).

predict_centiles() will work with this function, but it will likely throw a warning message. Therefore, predict_gen_gamma_gamlss() provides an alternative way to compute centiles from the model. It manually computes the centiles instead of relying on gamlss::centiles().

Examples